requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
idna==3.7
mysqlclient==2.2.4
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
